chore(gestures): move ui/pointer events to gestures
