+++ src/components/range/test/range.spec.ts
@@ -1,5 +1,5 @@
 import { Range } from '../range';
-import { mockConfig, mockRenderer, mockElementRef, mockHaptic } from '../../../util/mock-providers';
+import { mockConfig, MockDomController, mockRenderer, mockElementRef, mockHaptic } from '../../../util/mock-providers';
 import { Form } from '../../../util/form';
 
 
@@ -65,5 +65,5 @@ describe('Range', () => {
 
 function createRange(): Range {
   let form = new Form();
-  return new Range(form, mockHaptic(), null, mockConfig(), mockElementRef(), mockRenderer());
+  return new Range(form, mockHaptic(), null, mockConfig(), mockElementRef(), mockRenderer(), new MockDomController());
 }
